+++ src/utils/router/router.js
@@ -3,6 +3,10 @@ import ReactDOM from 'react-dom';
 import Router from 'react-router';
 import createBrowserHistory from 'history/lib/createBrowserHistory';
 
+let scrollToTop = () => {
+  window.scrollTo(0, 0);
+};
+
 /**
  * Exposes the history object to allow developers to redirect and control the
  * route state (see https://github.com/ReactJSTraining/history/tree/master/docs)
@@ -40,7 +44,7 @@ export function startRouter(routes, target = document.getElementById('app')) {
   // render the router into the DOM
   if (target) {
     ReactDOM.render((
-      <Router history={ history }>
+      <Router onUpdate={scrollToTop} history={ history }>
         { routes }
       </Router>
     ), target);